pyproject.toml
[tool.poetry]
name = "mixy"
version = "1.8.0"
description = ""
authors = ["Alexis Beaulieu <alexisbeaulieu97@gmail.com>"]
license = "MIT"
readme = "README.md"
homepage = "https://github.com/alexisbeaulieu97/Mixy"
repository = "https://github.com/alexisbeaulieu97/Mixy"

[tool.poetry.dependencies]
python = "^3.11"
jinja2 = "^3.1.2"
typer = {extras = ["all"], version = "^0.7.0"}
pyyaml = "^6.0"
pydantic = "^1.10.2"
requests = "^2.28.1"
gitpython = "^3.1.29"


[tool.poetry.group.dev.dependencies]
black = "^22.10.0"
mypy = "^0.991"
pylama = "^8.4.1"
python-semantic-release = "^7.32.2"
pytest = "^7.2.0"
pytest-cov = "^4.0.0"
pylint = "^2.15.9"
flake8 = "^6.0.0"
pyfakefs = "^5.0.0"
vulture = "^2.7"
pytest-mock = "^3.10.0"
semgrep = "^1.14.0"
bandit = {extras = ["toml"], version = "^1.7.4"}

[tool.poetry.scripts]
mixy = "mixy.cli:main"

[tool.semantic_release]
version_variable = ["mixy/__init__.py:__version__"]
version_toml = ["pyproject.toml:tool.poetry.version"]
version_pattern = ["VERSION:{version}"]
branch = "main"
upload_to_repository = true
upload_to_release = true
build_command = "pip install poetry && poetry build"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pylama]
skip = "*/__init__.py"
max_line_length = 120
